+++ Reservation/Migrations/ApplicationDBContextModelSnapshot.cs
@@ -24,11 +24,11 @@ namespace Reservation.Migrations
 
             modelBuilder.Entity("Reservation.Models.Equipment", b =>
                 {
-                    b.Property<long>("IdEquipment")
+                    b.Property<int>("IdEquipment")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("IdEquipment"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("IdEquipment"));
 
                     b.Property<string>("EquipmentName")
                         .IsRequired()
@@ -96,11 +96,11 @@ namespace Reservation.Migrations
 
             modelBuilder.Entity("Reservation.Models.Room", b =>
                 {
-                    b.Property<long>("IdRoom")
+                    b.Property<int>("IdRoom")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("IdRoom"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("IdRoom"));
 
                     b.Property<bool>("AirConditioning")
                         .HasColumnType("bit");
